Create To-Do List (JavaScript)
This simple to-do list project allows users to enter tasks in an input field and click the "Add" button to add them to the list. Each task is displayed as a list item. Users can click on a task to mark it as completed, which will apply a strikethrough effect to the task text.